Created a new class for CLIO and separated all CLIO from dealer class to put into the new class"
