+++ spec/dealer_spec.rb
@@ -2,54 +2,6 @@ require 'dealer'
 
 describe Dealer do
 
-	it "ask if user wants to hit or stop" do
-		output = StringIO.new
-		dealer = Dealer.new(output)
-		dealer.ask_hit_or_stop
-		expect(output.string).to eq("Do you want to hit or stop?\n")
-	end
-
-	it "gets input from the user" do
-		output = StringIO.new
-		input = StringIO.new("hit")
-		dealer = Dealer.new(output, input)
-		expect(dealer.get_hit_or_stop).to eq("hit")
-	end
-
-	it "returns score if user enters 'stop'" do
-		output = StringIO.new
-		dealer = Dealer.new(output)
-		dealer.return_score(10)
-		expect(output.string).to eq("Game over and your score is: 10\n")
-	end
-
-	it "displays '1 or 2 player?'" do
-		output = StringIO.new
-		dealer = Dealer.new(output)
-		dealer.ask_how_many_players
-    expect(output.string).to eq("1 or 2 player?\n")
-	end
-
-	it "displays an error message if user enters 0" do
-		output = StringIO.new
-		input = StringIO.new("0\n1")
-    dealer = Dealer.new(output, input)
-		dealer.get_num_of_players
-		expect(output.string).to eq("Please enter 1 or 2\n")
-	end
-
-  it "gets 1 from the user" do
-		output = StringIO.new
-		input = StringIO.new("1")
-    dealer = Dealer.new(output, input)
-		expect(dealer.get_num_of_players).to eq(1)
-	end
-
-	it "gets 2 from the user" do
-		output = StringIO.new
-		input = StringIO.new("2")
-		dealer = Dealer.new(output, input)
-		expect(dealer.get_num_of_players).to eq(2)
-	end
+	
 
 end
